Add basic parser, codegen, and executor
This commit does a few things. It firsts adds a parser so that we can
parse sycamore programs. It also adds the ability to compile the program
with some basic code generation to wasm and then we also add the ability
to run it by setting up an instance with the ability to println so that
we can see what our program actually does. With this we now have a very
basic programming language. It can't do much but it's a start!