Put run and build in lib.rs and start stdlib
The CLI is meant to just be the code to act as an entry to running
everything, but libsyc is meant to be all the code that's actually
taking in input parsing, doing codegen, and running. With this we
split the code out of main.rs and into lib.rs. This means anyone could
embed the sycamore runtime into their program if they so chose. We also
create our standard library finally and change the code such that we now
have a print and println function available. Future work will make it so
that we can have more flexibility in our codegen such that the stdlib
library knows how to generate the needed code if need be, but also so
that we don't have to change numbers by hand and make things error
prone.