Hoist codegen data types into the Codegen struct
In order to further simplify code generation and make it easier to
extend we hoist the various sections of a wasm module into the Codgen
struct as well things like our string literal table and function map.
This way we can do work to make it easier to let types implement a
future trait `Generate` so that they can make the code for them easily
by having access to all the necessary types in one easy struct.