Add Generate and StdLib traits and change Codegen
In order to clean things up a bit we add the Generate and StdLib traits.
Generate will be used for any code types we want to generate code so
just functions for now. With StdLib we *also* generate code since it has
the Generate Trait bound, but we also have it generate the function used
by wasmtime for it and how to add itself to the import section.

The design is by no means perfect but it's a great start. Hopefully this
will make things just a tad easier to work with in terms of extending
everything or adding more functionality.