Split Parser types into their own module
We were using types both in Codegen and the Parser. In order to separate
things better we split the types into their own module so that we can
use them across modules without confusing what the role of each type is.
We parse them into the type now and then use the type for codegen!