Clean up Statement codegen & make import_stdlib
This commit does two things, since function number is now tied to the
type number and we just repeat type sigs no matter what we can now use
that everywhere instead of having to figure it out and look up the type.
Especially when we can't inspect what types we already defined. Maybe we
can do an optimization pass here in the future. For now this works well
enough. With this we simplified some of the code in the codegen for the
Statement type. We also create an import_stdlib function so that we
import all of stdlib first and have one place to do all of those
imports with a somewhat stable order.