Add the ability to assign values in sycamore
With these changes number values can be assigned to a variable in the
form seen below:

my-variable <- 420;

We can't do anything else except for assigning a value, but this sets us
up to do more complicated things like referencing them in say a print
statement! We can then expand it to work with multiple types, do control
flow, do math etc. This is a start to unlocking more fun things! We
mostly just expand the codegen in `src/types.rs` for
`Statement::StateDefn` to setup all the locals we need in the function
and create a map so that we can make sure we assign to the right local
value!