Add `NameSection` to compiled wasm to debug code
This commit adds a new section when compiling code to have name output
when debugging. While it's possible to just work with numbers for
looking at what func is called or what local exists, it's in general
harder for humans to know what it references in the code generation
itself. With this when running code with the debug flag the wat output
will now contain the names of things instead of just the number to make
debugging a tad easier.