Change Option<Function> to Function in `Codegen`
This changes the type so that it's easier to work with by starting off
with a dummy function type and then overwriting it with the new correct
version for every single function when we start to actually make it.
This simplifies some of the code we work with in terms of reading it so
that we don't need to call map at the cost of a tiny negligible bit of
time in compilation.