Setup all string literals ahead of time
Rather than have the code be duplicated across multiple locations to
setup our string literals we instead set them all up in one spot. We
then change the table to a map of the string to the offset value in
memory. This also paves the way for string literals in assignment as
well so that we don't have to duplicate code.