Add AOT compilation & option to run compiled wasm
This commit compiles all of the code ahead of time for use in wasmtime.
This means on every startup all that has to be done is to read the file
in from disk or from memory if running and building the code. This way
we don't have to JIT everything all the time. This will be ideal as
programs get larger and larger. We also add a wasm flag so that we know
whether the path is a sycamore source file or not which will now know
how to either deserialize the file from disk or do the normal reading in
a sycamore file and running it!