Make initial implementation of WASI for Sycamore
This commit implements output in WASI rather than the custom
implementation we had done before which tied the output to the execution
engine with our custom function calls. This is a tad harder to implement
as the calls operate inside of the memory of the wasm, but the result is
we have a binary file that can run anywhere which is better than a file
that only runs locally.

The initial implementation only includes the `fd_write` call and
specifically just to stdout, but over time we'll change this to include
more of the `wasi_unstable` module and expand functionality of current
functions.

In order to make this universally compatible the AOT compilation
functionality was removed. We also went back to the normal offset
calculation at runtime for `StrLit`. This might change in the future,
but this is fine for now.