Implement CallerExt to reduce written code size
Rather than having lots of repeated code when accessing exports or the
data in the store by hand, we can instead implement functions on caller
itself to reduce the amount of written code that we have. For now this
is done with `get_memory` but could easily be exteneded as needed for
the type to make it easier to write functions for the standard library.