Remove stray comment for validating wasm
When we were first getting started with sycamore we were using
wasmparser to validate the code, but it turns out that now that we run
it within wasmtime that code is validated there so we only need
wasmparser to extract the context from the custom code section now. We
remove this comment so as to not be misleading.