Relicense all code under GPLv3 License
This commit changes the entire codebase to be GPLv3 licensed code.
Mainly I just don't want corporations to use this work while still being
open and available for others to use so long as they follow the license.
They can still use it, but most lawyers/corps are allergic to it. If
they do want to use it they can contact me to relicense it and pay good
money for it.

As I am the sole author right now I am also retroactively applying this
license to code in all previous commits in this repository as well.